pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "acrge-lite"
version = "0.1.0"
description = "Autonomous Cloud Reliability and Governance Engine Lite (ACRGE Lite)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [
  { name = "ACRGE Engineering" }
]
keywords = ["azure", "ai", "agents", "langgraph", "incident-management", "sre"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Typing :: Typed"
]
dependencies = [
  "azure-functions>=1.21.3,<2.0.0",
  "azure-identity>=1.20.0,<2.0.0",
  "azure-keyvault-secrets>=4.9.0,<5.0.0",
  "azure-appconfiguration>=1.7.1,<2.0.0",
  "azure-servicebus>=7.13.0,<8.0.0",
  "azure-storage-blob>=12.24.1,<13.0.0",
  "azure-cosmos>=4.9.0,<5.0.0",
  "azure-monitor-opentelemetry>=1.6.5,<2.0.0",
  "opentelemetry-api>=1.30.0,<2.0.0",
  "opentelemetry-sdk>=1.30.0,<2.0.0",
  "pydantic>=2.10.6,<3.0.0",
  "pydantic-settings>=2.7.1,<3.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "jinja2>=3.1.5,<4.0.0",
  "httpx>=0.28.1,<1.0.0",
  "tenacity>=9.0.0,<10.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "langgraph>=0.3.18,<0.4.0",
  "langchain-text-splitters>=0.3.6,<0.4.0",
  "redis>=5.2.1,<6.0.0",
  "jsonschema>=4.23.0,<5.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4,<9.0.0",
  "pytest-asyncio>=0.25.2,<0.26.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "mypy>=1.14.1,<2.0.0",
  "ruff>=0.8.6,<0.9.0",
  "bandit>=1.8.2,<2.0.0",
  "pip-audit>=2.8.0,<3.0.0",
  "types-redis>=4.6.0.20241004",
  "types-jsonschema>=4.23.0.20241208"
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "core", "services", "rag"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q --strict-markers --disable-warnings"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
namespace_packages = true
explicit_package_bases = true
exclude = ["infra/", "pipelines/"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["infra", "pipelines"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "W", "N", "ASYNC", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
